Update DimReduction.R
Inclusion of 'check_duplicates = FALSE' argument for Rtsne function. Default of TRUE cause failure if a dataset happened to have samples with duplicate counts, but we will ensure the function will complete, as online documentation could not supply a reasonable argument as to why duplicate samples should be excluded from calculation.